breaking change: rename data module into execution_data
