+++ mxops/execution/cli.py
@@ -8,7 +8,7 @@ import os
 from pathlib import Path
 from mxops.config.config import Config
 from mxops.data import path
-from mxops.data.data import ScenarioData, delete_scenario_data
+from mxops.data.execution_data import ScenarioData, delete_scenario_data
 
 from mxops.enums import parse_network_enum
 from mxops.execution.scene import execute_directory, execute_scene
